Updated handblebars for final iteration
Removed the event ID and RSVP'd user IDs from the event Info modals.
This was previously used as a tool for debugging within the client, and
is no longer needed.
Additionally, previously used rows in the same event Info modals have
been updated to appear more polished. E.g. 'Date' and 'Time' are no
longer on separate rows, and are instead displayed as " (Time), Date",
as would be seen in a real-life context.
Lastly, a new row has been added that displays the current number of
RSVP'd users.