Updated RSVP and UnRSVP buttons to toggle
Updated the relevant handlebars and RSVP events so that the event cards
will only contain a usable button for the user. I.e. An UnRSVP button
will populate on cards that the user is already RSVP'd to, and vice-
versa, an RSVP button will populate on cards that the user is not yet
RSVP'd to.

Note: There is still an undesirable behavior where, if a user either
RSVPs or UnRSVPs to an event from their "Owned Events" list, the app
will automatically update the event list to return the user to the
"See All Events" events list, rather than keeping the user on their
"Owned Events" list.
This isn't app-breakingly bad, but would be preferable to fix.